chore: analytics launch delay: level 1. DEV tracking enabled per default
